tidied up structure
