+++ workers/unity/Assets/BlankProject.asmdef
@@ -4,7 +4,6 @@
         "Improbable.Gdk.Core",
         "Improbable.Gdk.Generated",
         "Improbable.Gdk.Mobile",
-        "Improbable.Gdk.TransformSynchronization",
         "Improbable.Gdk.PlayerLifecycle",
         "Unity.Entities",
         "Unity.Entities.Hybrid",
@@ -13,5 +12,10 @@
     "optionalUnityReferences": [],
     "includePlatforms": [],
     "excludePlatforms": [],
-    "allowUnsafeCode": false
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": []
 }
