+++ workers/unity/Assets/BlankProject.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 2c296922cd82f43cda9a859d58e81bc9
-NativeFormatImporter:
+guid: f3f9c8690a2ca5d4b973a84329f67bdb
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
