+++ workers/unity/Assets/BlankProject/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f7d2ac63e1fe62c49ae8544cffca611c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
