+++ workers/unity/Assets/BlankProject/Scripts/Config/EntityTemplates.cs
@@ -1,0 +1,26 @@
+using Improbable;
+using Improbable.Gdk.Core;
+using Improbable.Gdk.PlayerLifecycle;
+
+namespace BlankProject.Scripts.Config
+{
+    public static class EntityTemplates
+    {
+        public static EntityTemplate CreatePlayerEntityTemplate(string workerId, byte[] serializedArguments)
+        {
+            var clientAttribute = EntityTemplate.GetWorkerAccessAttribute(workerId);
+            var serverAttribute = UnityGameLogicConnector.WorkerType;
+
+            var template = new EntityTemplate();
+            template.AddComponent(new Position.Snapshot(), clientAttribute);
+            template.AddComponent(new Metadata.Snapshot("Player"), serverAttribute);
+
+            PlayerLifecycleHelper.AddPlayerLifecycleComponents(template, workerId, serverAttribute);
+
+            template.SetReadAccess(UnityClientConnector.WorkerType, MobileClientWorkerConnector.WorkerType, serverAttribute);
+            template.SetComponentWriteAccess(EntityAcl.ComponentId, serverAttribute);
+
+            return template;
+        }
+    }
+}
