+++ workers/unity/Assets/BlankProject/Scripts/Config/EntityTemplates.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7eea20373a33460e8f43a1e9aa8b7ae6
+timeCreated: 1564671499
