+++ workers/unity/Assets/BlankProject/Scripts/Metrics.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7c08e5d2711e91b459dce73a74ba77b7
+guid: ec37d18266594f3498c9e3b75aa14d8f
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
