+++ workers/unity/Assets/BlankProject/Scripts/Metrics/MetricSendSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 34d2e6523462b3e40a136594cbeca260
+guid: e6ca0c64025d38643b5a2eabc3343146
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
