+++ workers/unity/Assets/BlankProject/Scripts/Workers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f6d9cd5b5b132644bbf5e90a1b01a461
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
