+++ workers/unity/Assets/BlankProject/Scripts/Workers/UnityGameLogicConnector.cs
@@ -1,7 +1,6 @@
-﻿using Improbable;
+﻿using BlankProject.Scripts.Config;
 using Improbable.Gdk.Core;
 using Improbable.Gdk.PlayerLifecycle;
-using Improbable.Gdk.TransformSynchronization;
 using Improbable.Worker.CInterop;
 using UnityEngine;
 
@@ -13,7 +12,7 @@ namespace BlankProject
 
         private async void Start()
         {
-            PlayerLifecycleConfig.CreatePlayerEntityTemplate = CreatePlayerEntityTemplate;
+            PlayerLifecycleConfig.CreatePlayerEntityTemplate = EntityTemplates.CreatePlayerEntityTemplate;
 
             IConnectionFlow flow;
             ConnectionParameters connectionParameters;
@@ -42,23 +41,6 @@ namespace BlankProject
         {
             Worker.World.GetOrCreateSystem<MetricSendSystem>();
             PlayerLifecycleHelper.AddServerSystems(Worker.World);
-        }
-
-        private static EntityTemplate CreatePlayerEntityTemplate(string workerId, byte[] serializedArguments)
-        {
-            var clientAttribute = EntityTemplate.GetWorkerAccessAttribute(workerId);
-            var serverAttribute = WorkerType;
-
-            var template = new EntityTemplate();
-            template.AddComponent(new Position.Snapshot(), clientAttribute);
-            template.AddComponent(new Metadata.Snapshot("Player"), serverAttribute);
-            TransformSynchronizationHelper.AddTransformSynchronizationComponents(template, clientAttribute);
-            PlayerLifecycleHelper.AddPlayerLifecycleComponents(template, workerId, serverAttribute);
-
-            template.SetReadAccess(UnityClientConnector.WorkerType, MobileClientWorkerConnector.WorkerType, serverAttribute);
-            template.SetComponentWriteAccess(EntityAcl.ComponentId, serverAttribute);
-
-            return template;
         }
     }
 }
