+++ workers/unity/Assets/EntityCache/Resources.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c7041935a219dbb4c980f4a793ee8044
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
