+++ workers/unity/Assets/Resources/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 177ee6d85ca1a2c4a90bb679f58dab39
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
