+++ workers/unity/Assets/Resources/Prefabs/Worker.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a0f5d80a30577ad4db859b0ab906373f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
