+++ workers/unity/Assets/Resources/Prefabs/Worker/ClientWorker.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 160a9a4e2c1d16a4a883f915c10498a0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
