+++ workers/unity/Assets/Resources/Prefabs/Worker/GameLogicWorker.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 43c846c58bf01564d8a10b8fa8dd9146
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
