+++ workers/unity/Assets/Resources/Prefabs/Worker/MobileClientWorker.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 52be6d27873dad947b8b64c50c8443fa
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
