--- workers/unity/Assets/Scripts/MetricSendSystem.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: cbabbfb47a604310b0cacf8ab17c1a6a
-timeCreated: 1541604290
